# server/src/modules/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMID RASTAR PSYCH — API",
    "version": "1.0.0",
    "description": "Fastify + PostgreSQL + Drizzle. Response format یکدست: { success, data, message } | { success:false, error:{code,message} }"
  },
  "servers": [{ "url": "/" }],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "sid" }
    }
  },
  "paths": {
...
    "/api/users/me": {
      "get": { "tags": ["users"], "summary": "پروفایل + آمار + کدهای من", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "{user, counts, coupons}" } } },
      "patch": { "tags": ["users"], "summary": "تغییر نام", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "user" } } }
    },
    "/api/users/me/change-password": { "post": { "tags": ["users"], "summary": "تغییر رمز", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "ok" } } } },
    "/api/users/me/avatar": { "post": { "tags": ["users"], "summary": "آپلود آواتار (multipart field=file) یا پاک کردن", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "{avatarUrl}" } } } },
...
    "/api/orders": {
      "post": {
        "tags": ["orders"],
        "summary": "ثبت سفارش از سبد (transaction + stock + coupon). قیمت سمت کلاینت trust نمی‌شود.",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerName": { "type": "string" },
                  "customerPhone": { "type": "string" },
                  "customerEmail": { "type": "string" },
                  "shippingAddress": { "type": "string" },
                  "city": { "type": "string" },
                  "postalCode": { "type": "string" },
                  "notes": { "type": "string" },
                  "couponId": { "type": "string", "description": "از /api/coupons/my" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "order" }, "400": { "description": "EMPTY_CART / INSUFFICIENT_STOCK / INVALID_COUPON ..." } }
      },
      "get": { "tags": ["orders"], "summary": "سفارش‌های من (paginated)", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "{items, pagination}" } } }
    },
    "/api/orders/{id}": { "get": { "tags": ["orders"], "summary": "جزئیات سفارش (فقط مالک)", "security": [{ "cookieAuth": [] }], "responses": { "200": { "description": "order+items+history" } } } },
...
  }
}